# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Frontline Prep API",
    "description": "Backend for the ISSB preparation platform: catalog, question bank, mock tests, and the ISSB simulation suite. Question generation runs locally -- no external AI service is called.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/me": {
      "patch": {
        "tags": [
          "me"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_api_v1_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/stats": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "My Stats",
        "operationId": "my_stats_api_v1_me_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/dashboard": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Dashboard",
        "description": "The whole signed-in landing page in one request.",
        "operationId": "dashboard_api_v1_me_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/physical": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Log Physical",
        "operationId": "log_physical_api_v1_me_physical_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhysicalLogIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhysicalLogOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Physical History",
        "operationId": "physical_history_api_v1_me_physical_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PhysicalLogOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/physical/progress": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Physical Progress",
        "description": "Latest metrics measured against the target program's standards.",
        "operationId": "physical_progress_api_v1_me_physical_progress_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "program_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Program Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhysicalProgressOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Users",
        "operationId": "list_users_api_v1_admin_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AdminUserOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update User",
        "description": "Role and status changes are super-admin only, and never self-applied.\n\nLetting an admin edit their own row is how an account locks itself out or\nquietly promotes itself; both need a second person.",
        "operationId": "update_user_api_v1_admin_users__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminUserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "target_service": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "status",
          "email_verified",
          "created_at"
        ],
        "title": "AdminUserOut"
      },
      "AdminUserUpdate": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "email_verified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Verified"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AdminUserUpdate"
      },
...
      "DashboardOut": {
        "properties": {
          "user": {
            "additionalProperties": true,
            "type": "object",
            "title": "User"
          },
          "stats": {
            "$ref": "#/components/schemas/UserStatsOut"
          },
          "recent_attempts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recent Attempts",
            "default": []
          },
          "due_revision": {
            "type": "integer",
            "title": "Due Revision",
            "default": 0
          },
          "resume_attempt_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Attempt Id"
          },
          "next_stage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Stage"
          },
          "announcements": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Announcements",
            "default": []
          },
          "recommended": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recommended",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "user",
          "stats"
        ],
        "title": "DashboardOut",
        "description": "The signed-in landing page, assembled server-side to keep it one call."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_AdminUserOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUserOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[AdminUserOut]"
      },
...
      "Page_PhysicalLogOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PhysicalLogOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[PhysicalLogOut]"
      },
...
      "PhysicalLogIn": {
        "properties": {
          "logged_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logged On"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 280
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PhysicalLogIn"
      },
      "PhysicalLogOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "logged_on": {
            "type": "string",
            "format": "date-time",
            "title": "Logged On"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics",
            "default": {}
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "logged_on"
        ],
        "title": "PhysicalLogOut"
      },
      "PhysicalProgressOut": {
        "properties": {
          "logs": {
            "items": {
              "$ref": "#/components/schemas/PhysicalLogOut"
            },
            "type": "array",
            "title": "Logs",
            "default": []
          },
          "standards": {
            "additionalProperties": true,
            "type": "object",
            "title": "Standards",
            "default": {}
          },
          "latest": {
            "additionalProperties": true,
            "type": "object",
            "title": "Latest",
            "default": {}
          },
          "gaps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Gaps",
            "default": []
          },
          "bmi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bmi"
          }
        },
        "type": "object",
        "title": "PhysicalProgressOut"
      },
...
      "ProfileUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 24
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 12
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "target_service": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "target_program_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Program Id"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 250.0,
                "minimum": 100.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 250.0,
                "minimum": 25.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "preferences": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferences"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProfileUpdate"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "student",
          "instructor",
          "admin",
          "super_admin"
        ],
        "title": "Role"
      },
      "ServiceCode": {
        "type": "string",
        "enum": [
          "army",
          "air_force",
          "navy",
          "common"
        ],
        "title": "ServiceCode"
      },
...
      "UserPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "target_service": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "target_program_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Program Id"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified"
          },
          "preferences": {
            "additionalProperties": true,
            "type": "object",
            "title": "Preferences",
            "default": {}
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "status",
          "email_verified",
          "created_at"
        ],
        "title": "UserPublic",
        "description": "The account payload returned alongside a token and from ``/auth/me``."
      },
      "UserStatsOut": {
        "properties": {
          "attempts_total": {
            "type": "integer",
            "title": "Attempts Total",
            "default": 0
          },
          "questions_answered": {
            "type": "integer",
            "title": "Questions Answered",
            "default": 0
          },
          "questions_correct": {
            "type": "integer",
            "title": "Questions Correct",
            "default": 0
          },
          "accuracy": {
            "type": "number",
            "title": "Accuracy",
            "default": 0.0
          },
          "study_seconds": {
            "type": "integer",
            "title": "Study Seconds",
            "default": 0
          },
          "current_streak": {
            "type": "integer",
            "title": "Current Streak",
            "default": 0
          },
          "longest_streak": {
            "type": "integer",
            "title": "Longest Streak",
            "default": 0
          },
          "last_active_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Active On"
          },
          "readiness": {
            "additionalProperties": true,
            "type": "object",
            "title": "Readiness",
            "default": {}
          },
          "topic_mastery": {
            "additionalProperties": true,
            "type": "object",
            "title": "Topic Mastery",
            "default": {}
          },
          "olq_profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Olq Profile",
            "default": {}
          }
        },
        "type": "object",
        "title": "UserStatsOut"
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "pending",
          "active",
          "suspended"
        ],
        "title": "UserStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT access token",
        "scheme": "bearer"
      }
    }
  }
}